+++ examples/data-sources/meraki_wireless_location_scanning_receiver/data-source.tf
@@ -1,0 +1,4 @@
+data "meraki_wireless_location_scanning_receiver" "example" {
+  id              = "12345678"
+  organization_id = "123456"
+}