+++ examples/data-sources/meraki_wireless_location_scanning_receivers/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_wireless_location_scanning_receivers" "example" {
+  organization_id = "123456"
+}