+++ examples/resources/meraki_wireless_location_scanning_receiver/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_wireless_location_scanning_receiver.example "<organization_id>,<id>"